requirements.txt
# Core dependencies
gradio>=4.0.0
plotly>=5.18.0
openai>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0
python-dotenv>=1.0.0
joblib>=1.3.0
lexical-diversity>=0.1.1

# Transformer dependencies
transformers>=4.35.0
torch>=2.1.0
accelerate>=0.24.0
peft>=0.6.0
bitsandbytes>=0.41.0

# Agent dependencies
langchain>=0.1.0
langchain-openai>=0.0.2

# RAG dependencies
faiss-cpu>=1.7.4

# API dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
